Forgot to take DM in to consideration. Also fixed 'bot added' message being red.

+++ commands/watermark.js
@@ -16,7 +16,7 @@ module.exports = {
 			opacity: 50,
 			url: null
 		};
-		msg.delete();
+		if (msg.channel.type !== "dm") msg.delete();
 
 		let pass = {url: null,pass: false},
 			imageUrl = new RegExp(/^(http|https):\/\/([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-])?(\.jpg|\.jpeg|\.png|\.webp)/, "i");
